requirements.txt
pandas
scikit-learn
paho-mqtt